rearrange function in and out of main script
