+++ convert.py
@@ -3,11 +3,129 @@
 
 import time  # Time module for calculating runtime of script
 import re  # module for regex
-from packages import *
+from packages import * # Import local packages
+from datetime import datetime
 
+
+timenow = datetime.now() # Get date and time into variable
 
 # Variable to calc and store how many lines in file being used as input
 num_lines = sum(1 for line in open('netscreen_config.txt'))
+
+class master():
+
+    succeeded = 0  # Number of lines converted
+    failed = 0  # Number of lines from input file not converted
+
+    default_app = {
+        "ANY": "any",
+        "BGP": "junos-bgp",
+        "CHARGEN": "junos-chargen",
+        "DHCP-Relay": "junos-dhcp-relay",
+        "DISCARD": "junos-discard",
+        "DNS": "junos-dns",
+        "ECHO": "junos-echo",
+        "FINGER": "junos-finger",
+        "FTP": "junos-ftp",
+        "GRE": "junos-gre",
+        "GTP": "junos-gtp",
+        "H.323": "junos-h323",
+        "HTTP": "junos-http",
+        "HTTP-EXT": "junos-http-ext",
+        "HTTPS": "junos-https",
+        "ICMP-ANY": "junos-icmp-all",
+        "IDENT": "junos-ident",
+        "IKE": "junos-ike",
+        "IKE-NAT": "junos-ike-nat",
+        "IMAP": "junos-imap",
+        "Internet Locator Service": "junos-internet-locator-service",
+        "IRC": "junos-irc",
+        "L2TP": "junos-l2tp",
+        "LDAP": "junos-ldap",
+        "LPR": "junos-lpr",
+        "MAIL": "junos-mail",
+        "MGCP-CA": "junos-mgcp-ca",
+        "MGCP-UA": "junos-mgcp-ua",
+        "MS-EXCHANGE-DIRECTORY": "junos-ms-rpc-msexchange",
+        "MS-EXCHANGE-INFO-STORE": "junos-ms-rpc-msexchange",
+        "MS-EXCHANGE-STORE": "junos-ms-rpc-msexchange",
+        "MS-IIS-COM": "junos-ms-rpc-iis-com",
+        "MS-RPC-ANY": "junos-ms-rpc-any",
+        "MS-RPC-EPM": "junos-ms-rpc-epm",
+        "MS-SQL": "junos-ms-sql",
+        "MSN": "junos-msn",
+        "NBDS": "junos-nbds",
+        "NBNAME": "junos-nbname",
+        "NFS": "junos-nfs",
+        "NNTP": "junos-nntp",
+        "NS Global": "junos-ns-global",
+        "NS Global PRO": "junos-ns-global-pro",
+        "NSM": "junos-nsm",
+        "NTP": "junos-ntp",
+        "OSPF": "junos-ospf",
+        "PC-Anywhere": "junos-pc-anywhere",
+        "PING": "junos-ping",
+        "POP3": "junos-pop3",
+        "PPTP": "junos-pptp",
+        "RADIUS": "junos-radius",
+        "Real Media": "junos-realaudio",
+        "RIP": "junos-rip",
+        "RSH": "junos-rsh",
+        "RTSP": "junos-rstp",
+        "SCCP": "junos-sccp",
+        "SCTP-ANY": "junos-sctp-any",
+        "SIP": "junos-sip",
+        "SMB": "junos-smb",
+        "SMTP": "junos-mail",
+        "SNMP": "udp_161",
+        "SQL Monitor": "junos-sql-monitor",
+        "SQL*Net V1": "junos-sqlnet-v1",
+        "SQL*Net V2": "junos-sqlnet-v2",
+        "SSH": "junos-ssh",
+        "SYSLOG": "junos-syslog",
+        "TALK": "junos-talk",
+        "TCP-ANY": "junos-tcp-any",
+        "TELNET": "junos-telnet",
+        "TFTP": "junos-tftp",
+        "UDP-ANY": "junos-udp-any",
+        "UUCP": "junos-uucp",
+        "VDO Live": "junos-vdo-live",
+        "VNC": "junos-vnc",
+        "WAIS": "junos-wais",
+        "WHOIS": "junos-whois",
+        "WINFRAME": "junos-winframe",
+        "X-WINDOWS": "junos-x-windows",
+        "YMSG": "junos-ymsg"
+    }
+
+    default_addr = {
+        "Any":  "any"
+    }
+
+    ## Dictionaries (Custom Service & Service Groups)
+    # For referencing when comes time to convert rules, contains Junos App's and App groups
+    service_ns_to_junos = {}  # key = ns_service_name, value = junos app
+
+    # Dictionary for ns multi service protocol to junos application set mapping
+    service_grp_to_app_set = {}  # key = ns service name, value = junos app set name
+
+    # Combine all 3 service dictionaries into 1 for ease of lookups in other function. This syntax only value in python >= 3.5
+    service_dicts = {} # Create empty and add above 3 service dicts later once populated
+
+    # Empty list to populate below to create address book names and bind to zones
+    list_of_zones = []
+
+    # Address name dictionary mapping.  Key = ns address name, value = junos address name
+    addresses_ns_to_junos = {}
+
+    # Address group to Junos address set mapping.  Key = ns address grp, value = junos address set
+    address_group_ns_to_junos_address_set = {}
+
+    # Address and address group dicts combines for lookups when building sets nested with sets
+    address_and_set_dicts = {}
+
+    # List containing src_zone, dst_zone and rule ID/name for backwards lookup for multiple dst/src/services in a rule
+    multi_rule_params = []
 
 
 def combine_dicts(*args):  # This dict combine syntax is ONLY valid in python >= 3.5
@@ -20,6 +138,22 @@ def combine_dicts(*args):  # This dict combine syntax is ONLY valid in python >=
     elif "address" in args:
         master.address_and_set_dicts = {**master.addresses_ns_to_junos, **master.address_group_ns_to_junos_address_set,
                                         **master.default_addr}
+
+
+timestamp = timenow.strftime(f'%Y%m%d_%H%M%S') # Change to useable variable to append to filenames
+
+def converted_config_output(line):  # Write Junos config to (OUTPUT)
+    converted = open(f'converted_{timestamp}.txt', "a")
+    converted.write(line + "\n")  # Write converted config and newline
+    master.succeeded += 1
+    converted.close()  # Close file
+
+
+def junk_file_output(line):  # Write lines not conforming to logic to a junk file for later review
+    junk = open(f'not_converted_{timestamp}.txt', "a")
+    junk.write(line)  # Write line to file for later review
+    master.failed += 1
+    junk.close()  # Close file
 
 
 def read_file():  # File to read Netscreen config from (INPUT) and then pass to dedicated functions based on regex
@@ -368,31 +502,6 @@ def multi_line_rule(line, type): # Multi src/dst/service rules
         junk_file_output(line)
 
 
-def sanity_check_naming(name): # Remove invalid characters from a string
-
-        # Set address_name var to same as address but replace anything in invalid_characters with "_" so works with Junos
-    invalid_characters = [" ", ".", "/", "\"", "\'", "\\", "!", "?", "[", "]", "{", "}", "|", "(", ")", "-", "+"]
-    
-    for chars in invalid_characters:
-        name = name.replace(chars, "_").lower()
-    
-    # Alpha numeric list for characters that Junos names are allowed to START with
-    alpha_num = ['a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z','0','1','2','3','4','5','6','7','8','9']
-
-    # Only if start (index 0 / 1st character) of name is NOT alphanumeric
-    if name[0].lower() not in alpha_num:
-        amended_name = "" # Blank var but populated 4 lines down so can be returned to function outside of loop
-        # loop to account for name starting with more than 1 non valid character such as ..dan or //dan
-        while name[0].lower() not in alpha_num:
-            amended_name = name[1:] # Slice from index 1 onwards, ie removal of 1st character
-            # print(f'address starting with invalid character = {amended_name}')    #   Debug invalid naming
-            name = amended_name # Change name to use the corrected string to stop loop when string starts alphanumeric
-        return amended_name
-
-    # Else, return name after earlier check and removal of non valid characters
-    else:
-        return name
-
 # last one is
 #print(master.address_and_set_dicts)
 #print(master.addresses_ns_to_junos)
@@ -405,5 +514,6 @@ def sanity_check_naming(name): # Remove invalid characters from a string
 
 
 if __name__ == "__main__":
+    start_time = time.time() # Used for overtime time of run
     read_file()  # The call to start the run of the functions
     print("Total Runtime:--- %s seconds ---" % (time.time() - start_time)) # Print out time it took to run this script from start to finish
